- Eclilpse 3.2.1で開いたらいろいろ怒られたので、黄色い波線が出ないように修正しました。 - 削除モードでチェックポイントも削除できるようにしました。 - ファイルを開くときに、ディレクトリが表示されない不具合を修正しました。
 この行以下は無視されます --

M    Editor$1.class
M    Editor$3.class
M    Editor$5.class
M    Editor$7.class
M    Editor$9.class
M    Editor$20.class
M    Editor$22.class
M    Editor$Mode.class
M    Editor$24.class
M    Editor$26.class
M    Editor$28.class
M    ShortSCurveRail.java
M    Stage.java
M    Editor.class
M    Editor$10.class
M    Editor$12.class
M    Editor$31.class
M    Editor$14.class
M    Editor$33.class
M    Editor$16.class
M    Editor.java
M    Rail.java
M    Editor$18.class
M    Editor$2.class
M    Util.class
M    Editor$4.class
M    Editor$6.class
M    Editor$8.class
M    Rail.class
M    Editor$21.class
M    Editor$23.class
M    Editor$25.class
M    Editor$27.class
M    Editor$29.class
M    NetKartFileFilter.class
M    Const.class
M    Util.java
M    NetKartFileFilter.java
M    Editor$11.class
M    Editor$30.class
M    Editor$13.class
M    Editor$32.class
M    Editor$15.class
M    Editor$17.class
M    Const.java
M    CheckPoint.java
M    Editor$19.class
M    Point.java

+++ ShortSCurveRail.java
@@ -24,7 +24,7 @@ public class ShortSCurveRail implements RailType {
 
 	public Shape getShape() {
 		final GeneralPath path = new GeneralPath();
-		final float fatness = (float) (Const.RAIL_WIDTH / 4);
+		final float fatness = Const.RAIL_WIDTH / 4;
 		path.moveTo(0, -Const.RAIL_WIDTH / 2);
 		path.curveTo(50 + fatness, -Const.RAIL_WIDTH / 2, 50 + fatness, 50 - Const.RAIL_WIDTH / 2, 100, 50 - Const.RAIL_WIDTH / 2);
 		path.moveTo(100, 50 + Const.RAIL_WIDTH / 2);
@@ -34,7 +34,7 @@ public class ShortSCurveRail implements RailType {
 
 	public Shape getFill() {
 		final GeneralPath path = new GeneralPath();
-		final float fatness = (float) (Const.RAIL_WIDTH / 4);
+		final float fatness = Const.RAIL_WIDTH / 4;
 		path.moveTo(0, -Const.RAIL_WIDTH / 2);
 		path.curveTo(50 + fatness, -Const.RAIL_WIDTH / 2, 50 + fatness, 50 - Const.RAIL_WIDTH / 2, 100, 50 - Const.RAIL_WIDTH / 2);
 		path.lineTo(100, 50 + Const.RAIL_WIDTH / 2);
